pyproject.toml
[project]
name = "medata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=4.0.6",
    "django-cors-headers>=3.11.0",
    "django-filter>=22.1",
    "django-import-export>=2.8.0",
    "djangorestframework>=3.13.1",
    "gunicorn>=22.0.0",
    "numpy>=1.22.3",
    "pandas>=1.4.2",
    "whitenoise>=6.9.0",
]
